LayoutTests/discovery/get-network-services-types-expected.txt

CONSOLE MESSAGE: line 12: okCDS() called...
CONSOLE MESSAGE: line 19: okCnn() called...
Tests navigator.getNetworkServices() works in happy case as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
